[move-ide] On-hover improvements (#19437)
## Description 

This PR improves on several instance of on-hover. 
1. Doc comments now preserves indentation

![image](https://github.com/user-attachments/assets/2380db0c-1619-4080-ba48-b438b1b7f0d5)
2. Clean up the way function signatures (in particular type params and
regular params) are displayed on-hover instead of always displaying them
on a single line:

![image](https://github.com/user-attachments/assets/17056202-4469-4ddb-a2be-a43564e7e0e2)
3. Use information about implicit modules/datatype when displaying type
information on-hover to abbreviate types by omitting package (for
implicit modules, e.g., `vector::singleton` instead of
`std::vector::singleton`) or both package and modules (for implicit
datatypes, e.g., `Option`, instead of `std::option::Option`)

![image](https://github.com/user-attachments/assets/ad639db4-a30a-4b70-852f-8614adca812f)

## Test plan 

All new and old tests must pass
