+++ external-crates/move/crates/move-analyzer/editors/code/package.json
@@ -5,7 +5,7 @@
   "publisher": "mysten",
   "icon": "images/move.png",
   "license": "Apache-2.0",
-  "version": "1.0.12",
+  "version": "1.0.13",
   "preview": true,
   "repository": {
     "url": "https://github.com/MystenLabs/sui.git",
