+++ external-crates/move/crates/move-analyzer/src/symbols.rs
@@ -90,10 +90,13 @@ use move_command_line_common::files::FileHash;
 use move_compiler::{
     command_line::compiler::{construct_pre_compiled_lib, FullyCompiledProgram},
     editions::{Edition, FeatureGate, Flavor},
-    expansion::ast::{self as E, AbilitySet, ModuleIdent, ModuleIdent_, Value, Value_, Visibility},
+    expansion::{
+        ast::{self as E, AbilitySet, ModuleIdent, ModuleIdent_, Value, Value_, Visibility},
+        name_validation::{IMPLICIT_STD_MEMBERS, IMPLICIT_STD_MODULES},
+    },
     linters::LintLevel,
     naming::ast::{DatatypeTypeParameter, StructFields, Type, TypeName_, Type_, VariantFields},
-    parser::ast as P,
+    parser::ast::{self as P},
     shared::{
         files::{FileId, MappedFiles},
         unique_map::UniqueMap,
@@ -109,6 +112,7 @@ use move_compiler::{
     unit_test::filter_test_members::UNIT_TEST_POISON_FUN_NAME,
     PASS_CFGIR, PASS_PARSER, PASS_TYPING,
 };
+use move_core_types::account_address::AccountAddress;
 use move_ir_types::location::*;
 use move_package::{
     compilation::{build_plan::BuildPlan, compiled_package::ModuleFormat},
@@ -118,7 +122,7 @@ use move_package::{
 use move_symbol_pool::Symbol;
 
 const MANIFEST_FILE_NAME: &str = "Move.toml";
-
+const STD_LIB_PKG_ADDRESS: &str = "0x1";
 type SourceFiles = BTreeMap<FileHash, (FileName, String, bool)>;
 
 /// Information about the compiled package and data structures
@@ -656,20 +660,35 @@ impl fmt::Display for DefInfo {
                 ret_type,
                 _,
             ) => {
-                let type_args_str = type_args_to_ide_string(type_args, /* verbose */ true);
+                const SINGLE_LINE_TYPE_ARGS_NUM: usize = 2;
+                // The strategy for displaying function signature is as follows:
+                // - if there are more than SINGLE_LINE_TYPE_ARGS_NUM type args,
+                //   they are displayed on separate lines
+                // - "regular" args are always displayed on separate lines, which
+                //   which is motivated by the fact that datatypes are displayed
+                //   in a fully-qualified form (i.e., with package and module name),
+                //   and that makes the function name already long and (likely)
+                //   the length of each individual type also long (modulo primitive
+                //   types of course, but I think we can live with that)
+                let type_args_str = type_args_to_ide_string(
+                    type_args,
+                    /* separate_lines */ type_args.len() > SINGLE_LINE_TYPE_ARGS_NUM,
+                    /* verbose */ true,
+                );
+                let args_str = typed_id_list_to_ide_string(
+                    arg_names, arg_types, '(', ')', /* separate_lines */ true,
+                    /* verbose */ true,
+                );
                 let ret_type_str = ret_type_to_ide_str(ret_type, /* verbose */ true);
                 write!(
                     f,
-                    "{}{}fun {}::{}{}({}){}",
+                    "{}{}fun {}{}{}{}{}",
                     visibility_to_ide_string(visibility),
                     fun_type_to_ide_string(fun_type),
-                    mod_ident_to_ide_string(mod_ident),
+                    mod_ident_to_ide_string(mod_ident, None, true),
                     name,
                     type_args_str,
-                    typed_id_list_to_ide_string(
-                        arg_names, arg_types, /* separate_lines */ false,
-                        /* verbose */ true
-                    ),
+                    args_str,
                     ret_type_str,
                 )
             }
@@ -689,9 +708,9 @@ impl fmt::Display for DefInfo {
                 if field_names.is_empty() {
                     write!(
                         f,
-                        "{}struct {}::{}{}{} {{}}",
+                        "{}struct {}{}{}{} {{}}",
                         visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident),
+                        mod_ident_to_ide_string(mod_ident, Some(name), true),
                         name,
                         type_args_str,
                         abilities_str,
@@ -699,15 +718,17 @@ impl fmt::Display for DefInfo {
                 } else {
                     write!(
                         f,
-                        "{}struct {}::{}{}{} {{\n{}\n}}",
+                        "{}struct {}{}{}{} {}",
                         visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident),
+                        mod_ident_to_ide_string(mod_ident, Some(name), true),
                         name,
                         type_args_str,
                         abilities_str,
                         typed_id_list_to_ide_string(
                             field_names,
                             field_types,
+                            '{',
+                            '}',
                             /* separate_lines */ true,
                             /* verbose */ true
                         ),
@@ -721,9 +742,9 @@ impl fmt::Display for DefInfo {
                 if variants.is_empty() {
                     write!(
                         f,
-                        "{}enum {}::{}{}{} {{}}",
+                        "{}enum {}{}{}{} {{}}",
                         visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident),
+                        mod_ident_to_ide_string(mod_ident, Some(name), true),
                         name,
                         type_args_str,
                         abilities_str,
@@ -731,9 +752,9 @@ impl fmt::Display for DefInfo {
                 } else {
                     write!(
                         f,
-                        "{}enum {}::{}{}{} {{\n{}\n}}",
+                        "{}enum {}{}{}{} {{\n{}\n}}",
                         visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident),
+                        mod_ident_to_ide_string(mod_ident, Some(name), true),
                         name,
                         type_args_str,
                         abilities_str,
@@ -745,30 +766,36 @@ impl fmt::Display for DefInfo {
                 if field_types.is_empty() {
                     write!(
                         f,
-                        "{}::{}::{}",
-                        mod_ident_to_ide_string(mod_ident),
+                        "{}{}::{}",
+                        mod_ident_to_ide_string(mod_ident, Some(enum_name), true),
                         enum_name,
                         name
                     )
                 } else if *positional {
                     write!(
                         f,
-                        "{}::{}::{}({})",
-                        mod_ident_to_ide_string(mod_ident),
+                        "{}{}::{}({})",
+                        mod_ident_to_ide_string(mod_ident, Some(enum_name), true),
                         enum_name,
                         name,
-                        type_list_to_ide_string(field_types, /* verbose */ true)
+                        type_list_to_ide_string(
+                            field_types,
+                            /* separate_lines */ false,
+                            /* verbose */ true
+                        )
                     )
                 } else {
                     write!(
                         f,
-                        "{}::{}::{}{{{}}}",
-                        mod_ident_to_ide_string(mod_ident),
+                        "{}{}::{}{}",
+                        mod_ident_to_ide_string(mod_ident, Some(enum_name), true),
                         enum_name,
                         name,
                         typed_id_list_to_ide_string(
                             field_names,
                             field_types,
+                            '{',
+                            '}',
                             /* separate_lines */ false,
                             /* verbose */ true,
                         ),
@@ -778,8 +805,8 @@ impl fmt::Display for DefInfo {
             Self::Field(mod_ident, struct_name, name, t, _) => {
                 write!(
                     f,
-                    "{}::{}\n{}: {}",
-                    mod_ident_to_ide_string(mod_ident),
+                    "{}{}\n{}: {}",
+                    mod_ident_to_ide_string(mod_ident, Some(struct_name), true),
                     struct_name,
                     name,
                     type_to_ide_string(t, /* verbose */ true)
@@ -916,11 +943,17 @@ fn visibility_to_ide_string(visibility: &Visibility) -> String {
     visibility_str
 }
 
-pub fn type_args_to_ide_string(type_args: &[Type], verbose: bool) -> String {
+pub fn type_args_to_ide_string(type_args: &[Type], separate_lines: bool, verbose: bool) -> String {
     let mut type_args_str = "".to_string();
     if !type_args.is_empty() {
         type_args_str.push('<');
-        type_args_str.push_str(&type_list_to_ide_string(type_args, verbose));
+        if separate_lines {
+            type_args_str.push('\n');
+        }
+        type_args_str.push_str(&type_list_to_ide_string(type_args, separate_lines, verbose));
+        if separate_lines {
+            type_args_str.push('\n');
+        }
         type_args_str.push('>');
     }
     type_args_str
@@ -939,10 +972,12 @@ fn datatype_type_args_to_ide_string(type_args: &[(Type, bool)], verbose: bool) -
 fn typed_id_list_to_ide_string(
     names: &[Name],
     types: &[Type],
+    list_start: char,
+    list_end: char,
     separate_lines: bool,
     verbose: bool,
 ) -> String {
-    names
+    let list = names
         .iter()
         .zip(types.iter())
         .map(|(n, t)| {
@@ -953,7 +988,12 @@ fn typed_id_list_to_ide_string(
             }
         })
         .collect::<Vec<_>>()
-        .join(if separate_lines { ",\n" } else { ", " })
+        .join(if separate_lines { ",\n" } else { ", " });
+    if separate_lines && !list.is_empty() {
+        format!("{}\n{}\n{}", list_start, list, list_end)
+    } else {
+        format!("{}{}{}", list_start, list, list_end)
+    }
 }
 
 pub fn type_to_ide_string(sp!(_, t): &Type, verbose: bool) -> String {
@@ -969,32 +1009,47 @@ pub fn type_to_ide_string(sp!(_, t): &Type, verbose: bool) -> String {
         }
         Type_::Apply(_, sp!(_, type_name), ss) => match type_name {
             TypeName_::Multiple(_) => {
-                format!("({})", type_list_to_ide_string(ss, verbose))
+                format!(
+                    "({})",
+                    type_list_to_ide_string(ss, /* separate_lines */ false, verbose)
+                )
             }
             TypeName_::Builtin(name) => {
                 if ss.is_empty() {
                     format!("{}", name)
                 } else {
-                    format!("{}<{}>", name, type_list_to_ide_string(ss, verbose))
+                    format!(
+                        "{}<{}>",
+                        name,
+                        type_list_to_ide_string(ss, /* separate_lines */ false, verbose)
+                    )
                 }
             }
-            TypeName_::ModuleType(sp!(_, module_ident), struct_name) => {
+            TypeName_::ModuleType(sp!(_, mod_ident), datatype_name) => {
                 let type_args = if ss.is_empty() {
                     "".to_string()
                 } else {
-                    format!("<{}>", type_list_to_ide_string(ss, verbose))
+                    format!(
+                        "<{}>",
+                        type_list_to_ide_string(ss, /* separate_lines */ false, verbose)
+                    )
                 };
                 if verbose {
-                    format!("{}::{}{}", module_ident, struct_name, type_args,)
+                    format!(
+                        "{}{}{}",
+                        mod_ident_to_ide_string(mod_ident, Some(&datatype_name.value()), true),
+                        datatype_name,
+                        type_args
+                    )
                 } else {
-                    struct_name.to_string()
+                    datatype_name.to_string()
                 }
             }
         },
         Type_::Fun(args, ret) => {
             format!(
                 "|{}| -> {}",
-                type_list_to_ide_string(args, verbose),
+                type_list_to_ide_string(args, /* separate_lines */ false, verbose),
                 type_to_ide_string(ret, verbose)
             )
         }
@@ -1004,12 +1059,18 @@ pub fn type_to_ide_string(sp!(_, t): &Type, verbose: bool) -> String {
     }
 }
 
-pub fn type_list_to_ide_string(types: &[Type], verbose: bool) -> String {
+pub fn type_list_to_ide_string(types: &[Type], separate_lines: bool, verbose: bool) -> String {
     types
         .iter()
-        .map(|t| type_to_ide_string(t, verbose))
+        .map(|t| {
+            if separate_lines {
+                format!("\t{}", type_to_ide_string(t, verbose))
+            } else {
+                type_to_ide_string(t, verbose)
+            }
+        })
         .collect::<Vec<_>>()
-        .join(", ")
+        .join(if separate_lines { ",\n" } else { ", " })
 }
 
 fn datatype_type_list_to_ide_string(types: &[(Type, bool)], verbose: bool) -> String {
@@ -1127,15 +1188,59 @@ fn ast_value_to_ide_string(sp!(_, val): &Value) -> String {
     }
 }
 
-pub fn mod_ident_to_ide_string(mod_ident: &E::ModuleIdent_) -> String {
+/// Creates a string representing a module ID, either on it's owne as in `pkg::module`
+/// or as part of a datatype or function type, in which it should be `pkg::module::`.
+/// If it's part of the datatype, name of the datatype is passed in `datatype_name_opt`.
+pub fn mod_ident_to_ide_string(
+    mod_ident: &ModuleIdent_,
+    datatype_name_opt: Option<&Symbol>,
+    is_access_chain_prefix: bool, // part of access chaing that should end with `::`
+) -> String {
     use E::Address as A;
+    // the module ID is to be a prefix to a data
+    let suffix = if is_access_chain_prefix { "::" } else { "" };
     match mod_ident.address {
-        A::Numerical {
-            name: None, value, ..
-        } => format!("{value}::{}", mod_ident.module).to_string(),
-        A::Numerical { name: Some(n), .. } | A::NamedUnassigned(n) => {
-            format!("{n}::{}", mod_ident.module).to_string()
+        A::Numerical { name, value, .. } => {
+            let pkg_name = match name {
+                Some(n) => n.to_string(),
+                None => value.to_string(),
+            };
+
+            let Ok(std_lib_pkg_address) = AccountAddress::from_hex_literal(STD_LIB_PKG_ADDRESS)
+            else {
+                // getting stdlib address did not work - use the whole thing
+                return format!("{pkg_name}::{}{}", mod_ident.module, suffix);
+            };
+            if value.value.into_inner() != std_lib_pkg_address {
+                // it's not a stdlib package - use the whole thing
+                return format!("{pkg_name}::{}{}", mod_ident.module, suffix);
+            }
+            // try stripping both package and module if this conversion
+            // is for a datatype, oherwise try only stripping package
+            if let Some(datatype_name) = datatype_name_opt {
+                if IMPLICIT_STD_MEMBERS.iter().any(
+                    |(implicit_mod_name, implicit_datatype_name, _)| {
+                        mod_ident.module.value() == *implicit_mod_name
+                            && datatype_name == implicit_datatype_name
+                    },
+                ) {
+                    // strip both package and module (whether its meant to be
+                    // part of access chain or not, if there is not module,
+                    // there should be no `::` at the end)
+                    return "".to_string();
+                }
+            }
+            if IMPLICIT_STD_MODULES
+                .iter()
+                .any(|implicit_mod_name| mod_ident.module.value() == *implicit_mod_name)
+            {
+                // strip package
+                return format!("{}{}", mod_ident.module.value(), suffix);
+            }
+            // stripping prefix didn't work - use the whole thing
+            format!("{pkg_name}::{}{}", mod_ident.module, suffix)
         }
+        A::NamedUnassigned(n) => format!("{n}::{}", mod_ident.module).to_string(),
     }
 }
 
@@ -2575,7 +2680,7 @@ fn get_mod_outer_defs(
         );
         def_info.insert(
             mod_defs.name_loc,
-            DefInfo::Module(mod_ident_to_ide_string(&ident), doc_comment),
+            DefInfo::Module(mod_ident_to_ide_string(&ident, None, false), doc_comment),
         );
     }
 
@@ -2712,7 +2817,7 @@ fn extract_doc_string(
     // Detect the two different types of docstrings
     if line_before.starts_with("///") {
         while let Some(stripped_line) = line_before.strip_prefix("///") {
-            doc_string = format!("{}\n{}", stripped_line.trim(), doc_string);
+            doc_string = format!("{}\n{}", stripped_line, doc_string);
             if iter == 0 {
                 break;
             }
@@ -2721,7 +2826,13 @@ fn extract_doc_string(
         }
     } else if line_before.ends_with("*/") {
         let mut doc_string_found = false;
-        line_before = file_lines[iter].strip_suffix("*/").unwrap_or("").trim();
+        // we need a line with preserved (whitespace) prefix so that
+        // we can trim other lines in the doc comment to the same prefix
+        let mut current_line = file_lines[iter].trim_end();
+        current_line = current_line.strip_suffix("*/").unwrap_or("");
+        let current_line_len = current_line.len();
+        line_before = current_line.trim();
+        let trimmed_len = current_line_len - line_before.len();
 
         // Loop condition is a safe guard.
         while !doc_string_found {
@@ -2745,7 +2856,36 @@ fn extract_doc_string(
             }
 
             iter -= 1;
-            line_before = file_lines[iter].trim();
+
+            // we need to trim the block comment line the same as
+            // the line containing its ending marker
+            current_line = file_lines[iter].trim_end();
+            let first_non_whitespace = current_line
+                .chars()
+                .position(|c| !c.is_whitespace())
+                .unwrap_or_default();
+            if first_non_whitespace < trimmed_len {
+                // There is not enough whitespace to trim but trim as much as you can.
+                // The reason likely is that the docsting is misformatted, for example:
+                // ```
+                //   /**
+                //     Properly formatted line
+                //     Another properly formatted line
+                // Misformatted line
+                //   */
+                // ```
+                //
+                // This will result in the following doc comment extracted:
+                //     Properly formatted line
+                //```
+                //   Properly formatted line
+                //   Another properly formatted line
+                //Misformatted line
+                //```
+                line_before = current_line.trim_start();
+            } else {
+                line_before = current_line[trimmed_len..].into();
+            }
         }
 
         // No doc_string found - return String::new();
