+++ external-crates/move/crates/move-analyzer/tests/docstring.ide
@@ -68,6 +68,18 @@
         {
           "use_line": 44,
           "use_ndx": 2
+        },
+        {
+          "use_line": 56,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 68,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 80,
+          "use_ndx": 0
         }
       ]
     }
