+++ external-crates/move/crates/move-analyzer/tests/dot_calls.exp
@@ -13,7 +13,9 @@ Use: 'foo', start: 25, end: 28
 Def: 'foo', line: 13, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::foo(s: &Move2024::M1::SomeStruct): u64
+public fun Move2024::M1::foo(
+	s: &Move2024::M1::SomeStruct
+): u64
 
 -- test 2 -------------------
 use line: 3, use_ndx: 2
@@ -31,7 +33,9 @@ Use: 'f1', start: 43, end: 45
 Def: 'foo', line: 13, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::foo(s: &Move2024::M1::SomeStruct): u64
+public fun Move2024::M1::foo(
+	s: &Move2024::M1::SomeStruct
+): u64
 
 -- test 4 -------------------
 use line: 4, use_ndx: 0
@@ -47,7 +51,9 @@ Use: 'foo', start: 33, end: 36
 Def: 'foo', line: 13, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::foo(s: &Move2024::M1::SomeStruct): u64
+public fun Move2024::M1::foo(
+	s: &Move2024::M1::SomeStruct
+): u64
 
 -- test 6 -------------------
 use line: 4, use_ndx: 2
@@ -65,7 +71,9 @@ Use: 'f2', start: 51, end: 53
 Def: 'foo', line: 13, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::foo(s: &Move2024::M1::SomeStruct): u64
+public fun Move2024::M1::foo(
+	s: &Move2024::M1::SomeStruct
+): u64
 
 -- test 8 -------------------
 use line: 27, use_ndx: 0
@@ -81,7 +89,10 @@ Use: 'bar', start: 22, end: 25
 Def: 'bar', line: 17, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::bar(s: &Move2024::M1::SomeStruct, v: u64): u64
+public fun Move2024::M1::bar(
+	s: &Move2024::M1::SomeStruct,
+	v: u64
+): u64
 
 -- test 10 -------------------
 use line: 27, use_ndx: 2
@@ -107,7 +118,10 @@ Use: 'f3', start: 54, end: 56
 Def: 'bar', line: 17, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::bar(s: &Move2024::M1::SomeStruct, v: u64): u64
+public fun Move2024::M1::bar(
+	s: &Move2024::M1::SomeStruct,
+	v: u64
+): u64
 
 -- test 13 -------------------
 use line: 30, use_ndx: 0
@@ -123,7 +137,10 @@ Use: 'bar', start: 20, end: 23
 Def: 'bar', line: 17, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::bar(s: &Move2024::M1::SomeStruct, v: u64): u64
+public fun Move2024::M1::bar(
+	s: &Move2024::M1::SomeStruct,
+	v: u64
+): u64
 
 -- test 15 -------------------
 use line: 30, use_ndx: 2
@@ -141,7 +158,10 @@ Use: 'f4', start: 43, end: 45
 Def: 'bar', line: 17, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::bar(s: &Move2024::M1::SomeStruct, v: u64): u64
+public fun Move2024::M1::bar(
+	s: &Move2024::M1::SomeStruct,
+	v: u64
+): u64
 
 -- test 17 -------------------
 use line: 34, use_ndx: 0
@@ -157,7 +177,9 @@ Use: 'f1', start: 28, end: 30
 Def: 'foo', line: 13, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::foo(s: &Move2024::M1::SomeStruct): u64
+public fun Move2024::M1::foo(
+	s: &Move2024::M1::SomeStruct
+): u64
 
 -- test 19 -------------------
 use line: 35, use_ndx: 0
@@ -173,7 +195,10 @@ Use: 'f3', start: 28, end: 30
 Def: 'bar', line: 17, def char: 15
 TypeDef: no info
 On Hover:
-public fun Move2024::M1::bar(s: &Move2024::M1::SomeStruct, v: u64): u64
+public fun Move2024::M1::bar(
+	s: &Move2024::M1::SomeStruct,
+	v: u64
+): u64
 
 -- test 21 -------------------
 use line: 35, use_ndx: 2
