+++ external-crates/move/crates/move-analyzer/tests/implicit_uses.exp
@@ -5,12 +5,12 @@ Use: 'Option', start: 13, end: 19
 Def: 'Option', line: 7, def char: 18
 TypeDef: 'Option', line: 7, char: 18
 On Hover:
-public struct std::option::Option<Element> has copy, drop, store {
+public struct Option<Element> has copy, drop, store {
 	vec: vector<Element>
 }
 
-Abstraction of a value that may or may not be present. Implemented with a vector of size
-zero or one because Move bytecode does not have ADTs.
+ Abstraction of a value that may or may not be present. Implemented with a vector of size
+ zero or one because Move bytecode does not have ADTs.
 
 
 -- test 1 -------------------
@@ -19,8 +19,8 @@ Use: 'option', start: 26, end: 32
 Def: 'option', line: 4, def char: 12
 TypeDef: no info
 On Hover:
-module std::option
+module option
 
-This module defines the Option type and its methods to represent and handle an optional value.
+ This module defines the Option type and its methods to represent and handle an optional value.
