+++ external-crates/move/crates/move-analyzer/tests/imports.exp
@@ -7,7 +7,7 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
 
 
 -- test 1 -------------------
