+++ external-crates/move/crates/move-analyzer/tests/mod_access.exp
@@ -7,7 +7,7 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
 
 
 -- test 1 -------------------
@@ -18,7 +18,7 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
 
 
 -- test 2 -------------------
@@ -29,7 +29,7 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
 
 
 -- test 3 -------------------
@@ -40,7 +40,7 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
 
 
 -- test 4 -------------------
@@ -51,7 +51,7 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
 
 
 -- test 5 -------------------
@@ -86,6 +86,6 @@ TypeDef: no info
 On Hover:
 module Symbols::M9
 
-A module doc comment
+ A module doc comment
