+++ external-crates/move/crates/move-analyzer/tests/symbols.exp
@@ -5,7 +5,7 @@ Use: 'SomeStruct', start: 11, end: 21
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -23,7 +23,9 @@ Use: 'unpack', start: 8, end: 14
 Def: 'unpack', line: 9, def char: 8
 TypeDef: no info
 On Hover:
-fun Symbols::M1::unpack(s: Symbols::M1::SomeStruct): u64
+fun Symbols::M1::unpack(
+	s: Symbols::M1::SomeStruct
+): u64
 
 -- test 3 -------------------
 use line: 10, use_ndx: 1
@@ -39,7 +41,7 @@ Use: 'SomeStruct', start: 18, end: 28
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -49,7 +51,7 @@ Use: 'SomeStruct', start: 12, end: 22
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -92,7 +94,7 @@ Use: 'SomeStruct', start: 16, end: 26
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -102,7 +104,7 @@ Use: 'SomeStruct', start: 18, end: 28
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -139,7 +141,9 @@ Use: 'some_other_struct', start: 21, end: 38
 Def: 'some_other_struct', line: 6, def char: 15
 TypeDef: no info
 On Hover:
-public fun Symbols::M2::some_other_struct(v: u64): Symbols::M2::SomeOtherStruct
+public fun Symbols::M2::some_other_struct(
+	v: u64
+): Symbols::M2::SomeOtherStruct
 
 -- test 16 -------------------
 use line: 26, use_ndx: 2
@@ -165,7 +169,9 @@ Use: 'acq', start: 8, end: 11
 Def: 'acq', line: 34, def char: 8
 TypeDef: no info
 On Hover:
-fun Symbols::M1::acq(uint: u64): u64
+fun Symbols::M1::acq(
+	uint: u64
+): u64
 
 -- test 19 -------------------
 use line: 41, use_ndx: 2
@@ -197,7 +203,7 @@ Use: 'SomeStruct', start: 15, end: 25
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -207,7 +213,7 @@ Use: 'SomeStruct', start: 27, end: 37
 Def: 'SomeStruct', line: 2, def char: 11
 TypeDef: 'SomeStruct', line: 2, char: 11
 On Hover:
-struct Symbols::M1::SomeStruct has drop, store, key {
+struct Symbols::M1::SomeStruct has drop, store {
 	some_field: u64
 }
 
@@ -258,7 +264,10 @@ Use: 'ret', start: 8, end: 11
 Def: 'ret', line: 61, def char: 8
 TypeDef: no info
 On Hover:
-fun Symbols::M1::ret(p1: bool, p2: u64): u64
+fun Symbols::M1::ret(
+	p1: bool,
+	p2: u64
+): u64
 
 -- test 30 -------------------
 use line: 64, use_ndx: 0
@@ -604,4 +613,94 @@ Def: 'SOME_CONST', line: 55, def char: 10
 TypeDef: no info
 On Hover:
 const Symbols::M5::SOME_CONST: u64 = 7
+
+-- test 11 -------------------
+use line: 58, use_ndx: 0
+Use: 'long_param_list', start: 15, end: 30
+Def: 'long_param_list', line: 57, def char: 15
+TypeDef: no info
+On Hover:
+public fun Symbols::M5::long_param_list(
+	foo: u64,
+	bar: u64,
+	baz: u64,
+	qux: u64
+)
+
+-- test 12 -------------------
+use line: 60, use_ndx: 0
+Use: 'short_type_param_list', start: 15, end: 36
+Def: 'short_type_param_list', line: 59, def char: 15
+TypeDef: no info
+On Hover:
+public fun Symbols::M5::short_type_param_list<TYPE1, TYPE2>()
+
+-- test 13 -------------------
+use line: 62, use_ndx: 0
+Use: 'long_type_param_list', start: 15, end: 35
+Def: 'long_type_param_list', line: 61, def char: 15
+TypeDef: no info
+On Hover:
+public fun Symbols::M5::long_type_param_list<
+	TYPE1,
+	TYPE2,
+	TYPE3
+>()
+
+-- test 14 -------------------
+use line: 64, use_ndx: 0
+Use: 'combined_short_type_param_list', start: 15, end: 45
+Def: 'combined_short_type_param_list', line: 63, def char: 15
+TypeDef: no info
+On Hover:
+public fun Symbols::M5::combined_short_type_param_list<TYPE1, TYPE2>(
+	foo: u64,
+	bar: u64,
+	baz: u64,
+	qux: u64
+)
+
+-- test 15 -------------------
+use line: 68, use_ndx: 0
+Use: 'combined_long_type_param_list', start: 15, end: 44
+Def: 'combined_long_type_param_list', line: 67, def char: 15
+TypeDef: no info
+On Hover:
+public fun Symbols::M5::combined_long_type_param_list<
+	TYPE1,
+	TYPE2,
+	TYPE3
+>(
+	foo: u64,
+	bar: u64,
+	baz: u64,
+	qux: u64
+)
+
+-- test 16 -------------------
+use line: 75, use_ndx: 2
+Use: 'extract', start: 37, end: 44
+Def: 'extract', line: 80, def char: 15
+TypeDef: no info
+On Hover:
+public fun option::extract<Element>(
+	t: &mut Option<Element>
+): Element
+
+ Convert a `some` option to a `none` by removing and returning the value stored inside `t`
+ Aborts if `t` does not hold a value
+
+
+-- test 17 -------------------
+use line: 78, use_ndx: 1
+Use: 'singleton', start: 21, end: 30
+Def: 'singleton', line: 64, def char: 15
+TypeDef: no info
+On Hover:
+public fun vector::singleton<Element>(
+	e: Element
+): vector<Element>
+
+ Return an vector of size one containing element `e`.
+
