+++ external-crates/move/crates/move-analyzer/tests/symbols.ide
@@ -295,6 +295,34 @@
         {
           "use_line": 56,
           "use_ndx": 0
+        },
+        {
+          "use_line": 58,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 60,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 62,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 64,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 68,
+          "use_ndx": 0
+        },
+        {
+          "use_line": 75,
+          "use_ndx": 2
+        },
+        {
+          "use_line": 78,
+          "use_ndx": 1
         }
       ]
     }
