+++ external-crates/move/crates/move-analyzer/tests/symbols/sources/M6.move
@@ -2,7 +2,7 @@ module Symbols::M6 {
 
     /// This is a documented struct
     /// With a multi-line docstring
-    struct DocumentedStruct has key, drop, store {
+    struct DocumentedStruct has drop, store {
         /// A documented field
         documented_field: u64,
     }
@@ -44,5 +44,39 @@ module Symbols::M6 {
     fun type_param_doc<T: copy + drop>(param: T): T {
         param
     }
+
+    /// A documented function with code block
+    /// (should preserved indentation in the code block)
+    ///
+    /// ```rust
+    /// fun foo() {
+    ///     42
+    /// }
+    /// ```
+    fun code_block_doc_slash() {}
+
+    /**
+       A documented function with code block
+       (should preserved indentation in the code block)
+
+       ```rust
+       fun foo() {
+           42
+       }
+       ```
+    */
+    fun code_block_doc_star() {}
+
+    /**
+      Misformatted docstring to have fewer whitespace in the body than
+      at the ending marker.
+
+
+  Beginning of this string should not disappear.
+
+Beginning of this string should not disappear either.
+
+    */
+    fun misformatted_docstring() {}
 
 }
