+++ external-crates/move/crates/move-compiler/src/expansion/translate.rs
@@ -13,6 +13,13 @@ use crate::{
         aliases::AliasSet,
         ast::{self as E, Address, Fields, ModuleIdent, ModuleIdent_, TargetKind},
         byte_string, hex_string,
+        name_validation::{
+            check_restricted_name_all_cases, check_valid_address_name,
+            check_valid_function_parameter_name, check_valid_local_name,
+            check_valid_module_member_alias, check_valid_module_member_name,
+            check_valid_type_parameter_name, valid_local_variable_name, ModuleMemberKind, NameCase,
+            IMPLICIT_STD_MEMBERS, IMPLICIT_STD_MODULES, IMPLICIT_SUI_MEMBERS, IMPLICIT_SUI_MODULES,
+        },
         path_expander::{
             access_result, Access, LegacyPathExpander, ModuleAccessResult, Move2024PathExpander,
             PathExpander,
@@ -327,35 +334,6 @@ fn compute_address_conflicts(
         .collect()
 }
 
-// Implicit aliases for the Move Stdlib:
-// use std::vector;
-// use std::option::{Self, Option};
-const IMPLICIT_STD_MODULES: &[Symbol] = &[symbol!("option"), symbol!("vector")];
-const IMPLICIT_STD_MEMBERS: &[(Symbol, Symbol, ModuleMemberKind)] = &[(
-    symbol!("option"),
-    symbol!("Option"),
-    ModuleMemberKind::Struct,
-)];
-
-// Implicit aliases for Sui mode:
-// use sui::object::{Self, ID, UID};
-// use sui::transfer;
-// use sui::tx_context::{Self, TxContext};
-const IMPLICIT_SUI_MODULES: &[Symbol] = &[
-    symbol!("object"),
-    symbol!("transfer"),
-    symbol!("tx_context"),
-];
-const IMPLICIT_SUI_MEMBERS: &[(Symbol, Symbol, ModuleMemberKind)] = &[
-    (symbol!("object"), symbol!("ID"), ModuleMemberKind::Struct),
-    (symbol!("object"), symbol!("UID"), ModuleMemberKind::Struct),
-    (
-        symbol!("tx_context"),
-        symbol!("TxContext"),
-        ModuleMemberKind::Struct,
-    ),
-];
-
 fn default_aliases(context: &mut Context) -> AliasMapBuilder {
     let current_package = context.current_package();
     let mut builder = context.new_alias_map_builder();
@@ -623,7 +601,7 @@ fn top_level_address_(
     suggest_declaration: bool,
     ln: P::LeadingNameAccess,
 ) -> Address {
-    let name_res = check_valid_address_name(context, &ln);
+    let name_res = check_valid_address_name(context.env, &ln);
     let sp!(loc, ln_) = ln;
     match ln_ {
         P::LeadingNameAccess_::AnonymousAddress(bytes) => {
@@ -661,7 +639,7 @@ pub(super) fn top_level_address_opt(
     context: &mut DefnContext,
     ln: P::LeadingNameAccess,
 ) -> Option<Address> {
-    let name_res = check_valid_address_name(context, &ln);
+    let name_res = check_valid_address_name(context.env, &ln);
     let named_address_mapping = context.named_address_mapping.as_ref().unwrap();
     let sp!(loc, ln_) = ln;
     match ln_ {
@@ -847,7 +825,7 @@ fn module_(
     assert!(context.address.is_none());
     assert!(address.is_none());
     set_module_address(context, &name, module_address);
-    let _ = check_restricted_name_all_cases(&mut context.defn_context, NameCase::Module, &name.0);
+    let _ = check_restricted_name_all_cases(context.defn_context.env, NameCase::Module, &name.0);
     if name.value().starts_with(|c| c == '_') {
         let msg = format!(
             "Invalid module name '{}'. Module names cannot start with '_'",
@@ -1461,7 +1439,7 @@ fn aliases_from_member(
 ) -> Option<P::ModuleMember> {
     macro_rules! check_name_and_add_implicit_alias {
         ($kind:expr, $name:expr) => {{
-            if let Some(n) = check_valid_module_member_name(context, $kind, $name) {
+            if let Some(n) = check_valid_module_member_name(&mut context.env(), $kind, $name) {
                 if let Err(loc) = acc.add_implicit_member_alias(
                     n.clone(),
                     current_module.clone(),
@@ -1598,7 +1576,7 @@ fn module_use(
     macro_rules! add_module_alias {
         ($ident:expr, $alias:expr) => {{
             if let Err(()) = check_restricted_name_all_cases(
-                &mut context.defn_context,
+                &mut context.defn_context.env,
                 NameCase::ModuleAlias,
                 &$alias,
             ) {
@@ -1678,7 +1656,8 @@ fn module_use(
 
                 let alias = alias_opt.unwrap_or(member);
 
-                let alias = match check_valid_module_member_alias(context, member_kind, alias) {
+                let alias = match check_valid_module_member_alias(context.env(), member_kind, alias)
+                {
                     None => continue,
                     Some(alias) => alias,
                 };
@@ -2259,7 +2238,7 @@ fn function_signature(
         .map(|(pmut, v, t)| (mutability(context, v.loc(), pmut), v, type_(context, t)))
         .collect::<Vec<_>>();
     for (_, v, _) in &parameters {
-        check_valid_function_parameter_name(context, is_macro, v)
+        check_valid_function_parameter_name(context.env(), is_macro, v)
     }
     let return_type = type_(context, pret_ty);
     E::FunctionSignature {
@@ -2307,7 +2286,7 @@ fn function_type_parameters(
         .into_iter()
         .map(|(name, constraints_vec)| {
             let constraints = ability_set(context, "constraint", constraints_vec);
-            let _ = check_valid_type_parameter_name(context, is_macro, &name);
+            let _ = check_valid_type_parameter_name(context.env(), is_macro, &name);
             (name, constraints)
         })
         .collect()
@@ -2320,7 +2299,7 @@ fn datatype_type_parameters(
     pty_params
         .into_iter()
         .map(|param| {
-            let _ = check_valid_type_parameter_name(context, None, &param.name);
+            let _ = check_valid_type_parameter_name(context.env(), None, &param.name);
             E::DatatypeTypeParameter {
                 is_phantom: param.is_phantom,
                 name: param.name,
@@ -3331,7 +3310,7 @@ fn bind(context: &mut Context, sp!(loc, pb_): P::Bind) -> Option<E::LValue> {
     let b_ = match pb_ {
         PB::Var(pmut, v) => {
             let emut = mutability(context, v.loc(), pmut);
-            check_valid_local_name(context, &v);
+            check_valid_local_name(context.env(), &v);
             EL::Var(Some(emut), sp(loc, E::ModuleAccess_::Name(v.0)), None)
         }
         PB::Unpack(ptn, pfields) => {
@@ -3573,389 +3552,4 @@ fn mutability(context: &mut Context, _loc: Loc, pmut: P::Mutability) -> E::Mutab
         // without let mut enabled, all locals are mutable and do not need the annotation
         None => E::Mutability::Either,
     }
-}
-
-//**************************************************************************************************
-// Valid names
-//**************************************************************************************************
-
-fn check_valid_address_name(
-    context: &mut DefnContext,
-    sp!(_, ln_): &P::LeadingNameAccess,
-) -> Result<(), ()> {
-    use P::LeadingNameAccess_ as LN;
-    match ln_ {
-        LN::AnonymousAddress(_) => Ok(()),
-        LN::GlobalAddress(n) | LN::Name(n) => {
-            check_restricted_name_all_cases(context, NameCase::Address, n)
-        }
-    }
-}
-
-fn valid_local_variable_name(s: Symbol) -> bool {
-    s.starts_with('_') || s.starts_with(|c: char| c.is_ascii_lowercase())
-}
-
-fn check_valid_function_parameter_name(context: &mut Context, is_macro: Option<Loc>, v: &Var) {
-    const SYNTAX_IDENTIFIER_NOTE: &str =
-        "'macro' parameters start with '$' to indicate that their arguments are not evaluated \
-        before the macro is expanded, meaning the entire expression is substituted. \
-        This is different from regular function parameters that are evaluated before the \
-        function is called.";
-    let is_syntax_identifier = v.is_syntax_identifier();
-    if let Some(macro_loc) = is_macro {
-        if !is_syntax_identifier && !v.is_underscore() {
-            let msg = format!(
-                "Invalid parameter name '{}'. '{}' parameter names must start with '$' (or must be '_')",
-                v, MACRO_MODIFIER,
-            );
-            let macro_msg = format!("Declared '{}' here", MACRO_MODIFIER);
-            let mut diag = diag!(
-                Declarations::InvalidName,
-                (v.loc(), msg),
-                (macro_loc, macro_msg),
-            );
-            diag.add_note(SYNTAX_IDENTIFIER_NOTE);
-            context.env().add_diag(diag);
-        }
-    } else if is_syntax_identifier {
-        let msg = format!(
-            "Invalid parameter name '{}'. Non-'{}' parameter names cannot start with '$'",
-            v, MACRO_MODIFIER,
-        );
-        let mut diag = diag!(Declarations::InvalidName, (v.loc(), msg));
-        diag.add_note(SYNTAX_IDENTIFIER_NOTE);
-        context.env().add_diag(diag);
-    } else if !is_valid_local_variable_name(v.value()) {
-        let msg = format!(
-            "Invalid parameter name '{}'. Local variable names must start with 'a'..'z', '_', \
-            or be a valid name quoted with backticks (`name`)",
-            v,
-        );
-        context
-            .env()
-            .add_diag(diag!(Declarations::InvalidName, (v.loc(), msg)));
-    }
-    let _ = check_restricted_name_all_cases(&mut context.defn_context, NameCase::Variable, &v.0);
-}
-
-fn check_valid_local_name(context: &mut Context, v: &Var) {
-    if !is_valid_local_variable_name(v.value()) {
-        let msg = format!(
-            "Invalid local name '{}'. Local variable names must start with 'a'..'z', '_', \
-            or be a valid name quoted with backticks (`name`)",
-            v,
-        );
-        context
-            .env()
-            .add_diag(diag!(Declarations::InvalidName, (v.loc(), msg)));
-    }
-    let _ = check_restricted_name_all_cases(&mut context.defn_context, NameCase::Variable, &v.0);
-}
-
-fn is_valid_local_variable_name(s: Symbol) -> bool {
-    Var::is_valid_name(s) && !Var::is_syntax_identifier_name(s)
-}
-
-#[derive(Copy, Clone, Debug)]
-pub enum ModuleMemberKind {
-    Constant,
-    Function,
-    Struct,
-    Enum,
-}
-
-impl ModuleMemberKind {
-    pub fn case(self) -> NameCase {
-        match self {
-            ModuleMemberKind::Constant => NameCase::Constant,
-            ModuleMemberKind::Function => NameCase::Function,
-            ModuleMemberKind::Struct => NameCase::Struct,
-            ModuleMemberKind::Enum => NameCase::Enum,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-pub enum NameCase {
-    Constant,
-    Function,
-    Struct,
-    Enum,
-    Module,
-    ModuleMemberAlias(ModuleMemberKind),
-    ModuleAlias,
-    Variable,
-    Address,
-    TypeParameter,
-}
-
-impl NameCase {
-    pub const fn name(&self) -> &'static str {
-        match self {
-            NameCase::Constant => "constant",
-            NameCase::Function => "function",
-            NameCase::Struct => "struct",
-            NameCase::Enum => "enum",
-            NameCase::Module => "module",
-            NameCase::ModuleMemberAlias(ModuleMemberKind::Function) => "function alias",
-            NameCase::ModuleMemberAlias(ModuleMemberKind::Constant) => "constant alias",
-            NameCase::ModuleMemberAlias(ModuleMemberKind::Struct) => "struct alias",
-            NameCase::ModuleMemberAlias(ModuleMemberKind::Enum) => "enum alias",
-            NameCase::ModuleAlias => "module alias",
-            NameCase::Variable => "variable",
-            NameCase::Address => "address",
-            NameCase::TypeParameter => "type parameter",
-        }
-    }
-}
-
-fn check_valid_module_member_name(
-    context: &mut Context,
-    member: ModuleMemberKind,
-    name: Name,
-) -> Option<Name> {
-    match check_valid_module_member_name_impl(context, member, &name, member.case()) {
-        Err(()) => None,
-        Ok(()) => Some(name),
-    }
-}
-
-fn check_valid_module_member_alias(
-    context: &mut Context,
-    member: ModuleMemberKind,
-    alias: Name,
-) -> Option<Name> {
-    match check_valid_module_member_name_impl(
-        context,
-        member,
-        &alias,
-        NameCase::ModuleMemberAlias(member),
-    ) {
-        Err(()) => None,
-        Ok(()) => Some(alias),
-    }
-}
-
-fn check_valid_module_member_name_impl(
-    context: &mut Context,
-    member: ModuleMemberKind,
-    n: &Name,
-    case: NameCase,
-) -> Result<(), ()> {
-    use ModuleMemberKind as M;
-    fn upper_first_letter(s: &str) -> String {
-        let mut chars = s.chars();
-        match chars.next() {
-            None => String::new(),
-            Some(c) => c.to_uppercase().collect::<String>() + chars.as_str(),
-        }
-    }
-    match member {
-        M::Function => {
-            if n.value.starts_with(|c| c == '_') {
-                let msg = format!(
-                    "Invalid {} name '{}'. {} names cannot start with '_'",
-                    case.name(),
-                    n,
-                    upper_first_letter(case.name()),
-                );
-                context
-                    .env()
-                    .add_diag(diag!(Declarations::InvalidName, (n.loc, msg)));
-                return Err(());
-            }
-        }
-        M::Constant | M::Struct | M::Enum => {
-            if !is_valid_datatype_or_constant_name(&n.value) {
-                let msg = format!(
-                    "Invalid {} name '{}'. {} names must start with 'A'..'Z'",
-                    case.name(),
-                    n,
-                    upper_first_letter(case.name()),
-                );
-                context
-                    .env()
-                    .add_diag(diag!(Declarations::InvalidName, (n.loc, msg)));
-                return Err(());
-            }
-        }
-    }
-
-    // TODO move these names to a more central place?
-    check_restricted_names(
-        context,
-        case,
-        n,
-        crate::naming::ast::BuiltinFunction_::all_names(),
-    )?;
-    check_restricted_names(
-        context,
-        case,
-        n,
-        crate::naming::ast::BuiltinTypeName_::all_names(),
-    )?;
-
-    // Restricting Self for now in the case where we ever have impls
-    // Otherwise, we could allow it
-    check_restricted_name_all_cases(&mut context.defn_context, case, n)?;
-
-    Ok(())
-}
-
-fn check_valid_type_parameter_name(
-    context: &mut Context,
-    is_macro: Option<Loc>,
-    n: &Name,
-) -> Result<(), ()> {
-    // TODO move these names to a more central place?
-    if n.value == symbol!("_") {
-        let diag = restricted_name_error(NameCase::TypeParameter, n.loc, "_");
-        context.env().add_diag(diag);
-        return Err(());
-    }
-
-    const SYNTAX_IDENTIFIER_NOTE: &str = "Type parameter names starting with '$' indicate that \
-        their arguments do not have to satisfy certain constraints before the macro is expanded, \
-        meaning types like '&mut u64' or '(bool, u8)' may be used as arguments.";
-
-    let is_syntax_ident = Var::is_syntax_identifier_name(n.value);
-    if let Some(macro_loc) = is_macro {
-        if !is_syntax_ident {
-            let msg = format!(
-                "Invalid type parameter name. \
-                '{} fun' type parameter names must start with '$'",
-                MACRO_MODIFIER
-            );
-            let macro_msg = format!("Declared '{}' here", MACRO_MODIFIER);
-            let mut diag = diag!(
-                Declarations::InvalidName,
-                (n.loc, msg),
-                (macro_loc, macro_msg),
-            );
-            diag.add_note(SYNTAX_IDENTIFIER_NOTE);
-            context.env().add_diag(diag);
-        } else {
-            let next_char = n.value.chars().nth(1).unwrap();
-            if !next_char.is_ascii_alphabetic() {
-                let msg = format!(
-                    "Invalid type parameter name '{}'. \
-                    Following the '$', the '{} fun' type parameter must be have a valid type \
-                    parameter name starting with a letter 'a'..'z' or 'A'..'Z'",
-                    n, MACRO_MODIFIER
-                );
-                let mut diag = diag!(Declarations::InvalidName, (n.loc, msg));
-                diag.add_note(SYNTAX_IDENTIFIER_NOTE);
-                context.env().add_diag(diag);
-            }
-        }
-    } else if is_syntax_ident {
-        let msg = format!(
-            "Invalid type parameter name. \
-                Only '{} fun' type parameter names cat start with '$'",
-            MACRO_MODIFIER
-        );
-        let mut diag = diag!(Declarations::InvalidName, (n.loc, msg));
-        diag.add_note(SYNTAX_IDENTIFIER_NOTE);
-        context.env().add_diag(diag);
-    }
-
-    // TODO move these names to a more central place?
-    check_restricted_names(
-        context,
-        NameCase::TypeParameter,
-        n,
-        crate::naming::ast::BuiltinFunction_::all_names(),
-    )?;
-    check_restricted_names(
-        context,
-        NameCase::TypeParameter,
-        n,
-        crate::naming::ast::BuiltinTypeName_::all_names(),
-    )?;
-
-    check_restricted_name_all_cases(&mut context.defn_context, NameCase::TypeParameter, n)
-}
-
-pub fn is_valid_datatype_or_constant_name(s: &str) -> bool {
-    s.starts_with(|c: char| c.is_ascii_uppercase())
-}
-
-// Checks for a restricted name in any decl case
-// Self and vector are not allowed
-fn check_restricted_name_all_cases(
-    context: &mut DefnContext,
-    case: NameCase,
-    n: &Name,
-) -> Result<(), ()> {
-    match case {
-        NameCase::Constant
-        | NameCase::Function
-        | NameCase::Struct
-        | NameCase::Enum
-        | NameCase::Module
-        | NameCase::ModuleMemberAlias(_)
-        | NameCase::ModuleAlias
-        | NameCase::Address => {
-            if Var::is_syntax_identifier_name(n.value) {
-                let msg = format!(
-                    "Invalid {} name '{}'. Identifiers starting with '$' can be used only for \
-                    parameters and type paramters",
-                    case.name(),
-                    n,
-                );
-                context
-                    .env
-                    .add_diag(diag!(Declarations::InvalidName, (n.loc, msg)));
-                return Err(());
-            }
-        }
-        NameCase::Variable | NameCase::TypeParameter => (),
-    }
-
-    let n_str = n.value.as_str();
-    let can_be_vector = matches!(case, NameCase::Module | NameCase::ModuleAlias);
-    if n_str == ModuleName::SELF_NAME
-        || (!can_be_vector && n_str == crate::naming::ast::BuiltinTypeName_::VECTOR)
-    {
-        context
-            .env
-            .add_diag(restricted_name_error(case, n.loc, n_str));
-        Err(())
-    } else {
-        Ok(())
-    }
-}
-
-fn check_restricted_names(
-    context: &mut Context,
-    case: NameCase,
-    sp!(loc, n_): &Name,
-    all_names: &BTreeSet<Symbol>,
-) -> Result<(), ()> {
-    if all_names.contains(n_) {
-        context
-            .env()
-            .add_diag(restricted_name_error(case, *loc, n_));
-        Err(())
-    } else {
-        Ok(())
-    }
-}
-
-fn restricted_name_error(case: NameCase, loc: Loc, restricted: &str) -> Diagnostic {
-    let a_or_an = match case.name().chars().next().unwrap() {
-        // TODO this is not exhaustive to the indefinite article rules in English
-        // but 'case' is never user generated, so it should be okay for a while/forever...
-        'a' | 'e' | 'i' | 'o' | 'u' => "an",
-        _ => "a",
-    };
-    let msg = format!(
-        "Invalid {case} name '{restricted}'. '{restricted}' is restricted and cannot be used to \
-         name {a_or_an} {case}",
-        a_or_an = a_or_an,
-        case = case.name(),
-        restricted = restricted,
-    );
-    diag!(NameResolution::ReservedName, (loc, msg))
 }
